@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 import org.sid.ebankingbackend.entities.BankAccount;
 import org.sid.ebankingbackend.enums.OperationType;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.Date;
 
 @Data
